refactor: Move front containers to their own subfolder
